Forge 1.17 - Early Untested
Port to Forge 1.17, kinda.

I can get a server for 1.17.1 running (and this mod runs fine on that) but client Forge 1.17.1 refuses to start for me.
All support threads related to the various errors are the regular forge 'helpers' telling people off and accusing them of using pirate launchers for no good reason, y'know, instead of providing support, so who knows how long it'll be before a solution emerges.